+++ src/main/features/claim/routes/eligibility/claimant-address.ts
@@ -7,7 +7,7 @@ import { Form } from 'forms/form'
 import User from 'idam/user'
 import { DraftService } from 'services/draftService'
 import { FormValidator } from 'forms/validation/formValidator'
-import { Eligibility } from 'drafts/models/eligibility/Eligibility'
+import { Eligibility } from 'claim/form/models/eligibility/eligibility'
 import { YesNoOption } from 'models/yesNoOption'
 import { NotEligibleReason } from 'claim/helpers/eligibility/notEligibleReason'
 import { ValidationGroups } from 'claim/helpers/eligibility/validationGroups'
