+++ src/main/features/claim/routes/eligibility/defendant-address.ts
@@ -6,7 +6,7 @@ import { ErrorHandling } from 'common/errorHandling'
 import { Form } from 'forms/form'
 import { YesNoOption } from 'models/yesNoOption'
 import { FormValidator } from 'app/forms/validation/formValidator'
-import { Eligibility } from 'app/drafts/models/eligibility/Eligibility'
+import { Eligibility } from 'claim/form/models/eligibility/eligibility'
 import User from 'idam/user'
 import { DraftService } from 'services/draftService'
 import { ValidationGroups } from 'claim/helpers/eligibility/validationGroups'
