+++ src/main/features/claim/routes/eligibility/government-department.ts
@@ -5,7 +5,7 @@ import { Paths } from 'claim/paths'
 import { ErrorHandling } from 'common/errorHandling'
 import { Form } from 'forms/form'
 import { FormValidator } from 'forms/validation/formValidator'
-import { Eligibility } from 'drafts/models/eligibility/eligibility'
+import { Eligibility } from 'claim/form/models/eligibility/eligibility'
 import { YesNoOption } from 'models/yesNoOption'
 import User from 'idam/user'
 import { DraftService } from 'services/draftService'
