+++ src/main/features/claim/routes/eligibility/help-with-fees.ts
@@ -6,7 +6,7 @@ import { ErrorHandling } from 'common/errorHandling'
 import { Form } from 'forms/form'
 import { DraftService } from 'services/draftService'
 import User from 'idam/user'
-import { Eligibility } from 'drafts/models/eligibility/Eligibility'
+import { Eligibility } from 'claim/form/models/eligibility/eligibility'
 import { FormValidator } from 'forms/validation/formValidator'
 import { YesNoOption } from 'models/yesNoOption'
 import { ValidationGroups } from 'claim/helpers/eligibility/validationGroups'
