+++ src/test/features/claim/routes/eligibility/over-18.ts
@@ -12,7 +12,7 @@ import { app } from '../../../../../main/app'
 
 import * as idamServiceMock from '../../../../http-mocks/idam'
 import * as draftStoreServiceMock from '../../../../http-mocks/draft-store'
-import { ClaimValue } from 'drafts/models/eligibility/claimValue'
+import { ClaimValue } from 'claim/form/models/eligibility/claimValue'
 import { NotEligibleReason } from 'claim/helpers/eligibility/notEligibleReason'
 
 const cookieName: string = config.get<string>('session.cookieName')
